requirements.txt
pandas
numpy
scikit-learn
scipy
scikit-surprise
PyYAML
